move state hash calculation to OperationExecutionRecord
