+++ libraries/rush-lib/src/logic/operations/Operation.ts
@@ -5,10 +5,6 @@ import type { RushConfigurationProject } from '../../api/RushConfigurationProjec
 import type { IPhase } from '../../api/CommandLineConfiguration';
 import type { IOperationRunner } from './IOperationRunner';
 import type { IOperationSettings } from '../../api/RushProjectConfiguration';
-import { RushConstants } from '../RushConstants';
-import * as crypto from 'crypto';
-import type { IInputsSnapshot } from '../incremental/InputsSnapshot';
-import type { BuildCacheConfiguration } from '../../api/BuildCacheConfiguration';
 
 /**
  * Options for constructing a new Operation.
@@ -153,84 +149,5 @@ export class Operation {
     // Cast internally to avoid adding the overhead of getters
     (this.dependencies as Set<Operation>).delete(dependency);
     (dependency.consumers as Set<Operation>).delete(this);
-  }
-
-  public get stateHash(): string {
-    if (!this._stateHash) {
-      throw new Error(
-        'Operation state hash is not calculated yet, you must call `calculateStateHash` first.'
-      );
-    }
-    return this._stateHash;
-  }
-
-  public calculateStateHash(options: {
-    inputsSnapshot: IInputsSnapshot;
-    buildCacheConfiguration: BuildCacheConfiguration;
-  }): string {
-    if (this._stateHash) {
-      return this._stateHash;
-    }
-    const { inputsSnapshot, buildCacheConfiguration } = options;
-    const { cacheHashSalt } = buildCacheConfiguration;
-    // This redefinition is necessary due to limitations in TypeScript's control flow analysis, due to the nested closure.
-    const definitelyDefinedInputsSnapshot: IInputsSnapshot = inputsSnapshot;
-
-    // Build cache hashes are computed up front to ensure stability and to catch configuration errors early.
-    function getOrCreateOperationHash(operation: Operation): string {
-      if (operation._stateHash) {
-        return operation._stateHash;
-      }
-      // Examples of data in the config hash:
-      // - CLI parameters (ShellOperationRunner)
-      const configHash: string | undefined = operation.runner?.getConfigHash();
-
-      const { associatedProject, associatedPhase } = operation;
-      // Examples of data in the local state hash:
-      // - Environment variables specified in `dependsOnEnvVars`
-      // - Git hashes of tracked files in the associated project
-      // - Git hash of the shrinkwrap file for the project
-      // - Git hashes of any files specified in `dependsOnAdditionalFiles` (must not be associated with a project)
-      const localStateHash: string | undefined =
-        associatedProject &&
-        definitelyDefinedInputsSnapshot.getOperationOwnStateHash(associatedProject, associatedPhase?.name);
-
-      // The final state hashes of operation dependencies are factored into the hash to ensure that any
-      // state changes in dependencies will invalidate the cache.
-      const dependencyHashes: string[] = Array.from(operation.dependencies, getDependencyHash).sort();
-
-      const hasher: crypto.Hash = crypto.createHash('sha1');
-      // This property is used to force cache bust when version changes, e.g. when fixing bugs in the content
-      // of the build cache.
-      hasher.update(`${RushConstants.buildCacheVersion}`);
-
-      if (cacheHashSalt !== undefined) {
-        // This allows repository owners to force a cache bust by changing the salt.
-        // A common use case is to invalidate the cache when adding/removing/updating rush plugins that alter the build output.
-        hasher.update(cacheHashSalt);
-      }
-
-      for (const dependencyHash of dependencyHashes) {
-        hasher.update(dependencyHash);
-      }
-
-      if (localStateHash) {
-        hasher.update(`${RushConstants.hashDelimiter}${localStateHash}`);
-      }
-
-      if (configHash) {
-        hasher.update(`${RushConstants.hashDelimiter}${configHash}`);
-      }
-
-      const hash: string = hasher.digest('hex');
-      operation._stateHash = hash;
-      return hash;
-    }
-
-    function getDependencyHash(operation: Operation): string {
-      return `${RushConstants.hashDelimiter}${operation.name}=${getOrCreateOperationHash(operation)}`;
-    }
-
-    return getOrCreateOperationHash(this);
   }
 }
